Support finding schemas using the dist directory
The contents of the dist directory is what is needed, so change the
GOVUK_CONTENT_SCHEMAS_PATH to refer to that instead. One of the
advantages of this is that applications using this library can use it in
development, and easily in production, where only the dist directory is
available.

This will require anything using this library to load schemas to change
the configuration of the environmental variable.