Make merge & validating an item faster
`merge_and_validate` currently uses the `payload` method which means
the generated random content is validated twice: once after the first
random generation and a second time after merging in the hash provided
by the user.

This refactors that process to makes sure we only validate the
resulting hash once. Only if the result turns out to be invalid, we do
a validation on the content item so that we know if the error is caused
by the random-generation code or the user.

The tests haven't changed because the behaviour is still the same.