Add method which excludes fields when generating a random example
The existing `merge_and_validate` method allows you to generate an
example with some overridden fields, but it does not allow you to
_remove_ optional fields from the payload.

This commit adds a `customise_and_validate` method which allows you to
do both. This lets you write a test which uses an example which is
missing a specific field such as `withdrawn_notice`, but with all the
other fields randomized.