Switch the order of payload and errors in error output
The JSON payload is quite big and means you need to scroll up to try
and find the actual errors, this can be a pain when multiple failures
have occurred. Switching the order should make iot easier to see the
errors without the need to scroll (for the last failure at least).