Support generating oneOf in objects
Currently if an object contains a `oneOf` property, the parent
`requires` and `properties` get ignored and only the contents of the
chosen `oneOf` object are used to generate a value.

This changes the way we generate objects which include a `oneOf`
property to take into account both the parent properties as well as the
`oneOf` properties.

This was necessary to support
https://github.com/alphagov/govuk-content-schemas/pull/910 which uses
`oneOf` to describe different required fields with different enum cases
which is not something we've done before on the content schemas.