Make random_example_spec.rb tests time-safe
I noticed that the test would occasionally fail because the time
would be different between the first run and second run of
`GovukSchemas::RandomExample.new(schema: schema).payload`

The `public_updated_at` might be 16:06:03 for one and 16:06:04 for
the other. This is just a natural consequence of the time it takes
to execute the code, and is not an indicator of the randomness not
being deterministic, which is what the test is testing.