Rename 'random.rb' => 'random_content_generator.rb'
This is to avoid clashing with the ruby standard lib
'Random' module, which we'll be using in the next
commit.

At the same time, I've renamed 'random_item_generator.rb' to
'random_schema_generator.rb', which is more reflective of its
purpose and helps to distinguish between the two classes.