Use ruby stdlib 'Random' for localised deterministic randomness
Prior to this commit, we relied on the global random seed set via
'srand', insofar as if somebody set 'srand', we'd return the same
result every time.
However, this could be surprising behaviour, as devs calling
RandomExample would expect a random example every time it is
called. We decided in an earlier [discussion] that it would be
better to pass an explicit 'seed' parameter if the dev wants
consistent results.

Whilst we could have used this 'seed' parameter to call
'srand(seed)', the behaviour becomes non-deterministic because
it changes the nature of the global environment. I.e. if the dev
is relying on setting 'srand' for some other part of their test
suite, we would now have skewed it by calling 'srand' again.
Hence the need to move to a sandbox/isolated way of setting
seeded random behaviour.

We used 'Random.new(seed)' and called all randomness-inducing
methods on that returned object, i.e. 'rand => @random.rand',
and 'arr.sample => arr.sample(random: @random)'.

[discussion]: https://github.com/alphagov/govuk_schemas/pull/56/files\#r458182188