Add reusable validator module
At the moment, applications that use Minitest still use the
govuk-content-schema-test-helpers gem.

As this has been deprecated, we need to add functionality to ensure that
applications that use Minitest can validate objects against the relevant
schema.

As a first step, i've pulled logic around validating a payload against
a schema and error construction into a reusable module. This allows
reuse in the assertion matchers that will be used for applications
that use Minitest in a subsequent commit.