Add github actions ci config and remove jenkinsfile
As per RFC-123 we're moving away from Jenkins for testing and publishing our gems via github actions.
Note that in the jenkinsfile we were using a rubyLintDiff: false option. This setting was actually deprecated in govuk_jenkinslib so has no effect here, and has not been ported to the github actions ci.yml