Make validator error message for readable and test with regexs
This updates the error message to use the name of the schema
in the error message rather than the UID.

It also updates the tests to test the error message that we have control
over specifically.

As we do not control the which properties are missing based on the
in this repo, we now just test that the gsubbing of the error messages
works as expected. And don't specifically test which properties are
missing.