Ensure the validator can handle the payload being JSON
At the moment, there's some disparity between how applications that
user Minitest and applications that use RSpec pass their payloads.

RSpec applications pass the payload as JSON and Minitest applications
pass it as a Hash.

Currently, the payload is being converted to JSON before being validated
by `JSON.fully_validate`. If the object is already JSON then .to_json is
called again. This means when JSON.parse is called on the object it does
not convert it back to a hash.

This commit updates the validator to ensure that the payload is
converted to JSON if the object passed is not a string.