Configurable content_schema_dir
This allows apps that use this gem to configuring the content schema
directory and not have to rely on an env var.

The primary motivation for this feature is in moving content schemas to
Publishing API [1] and allowing Publishing API to be able to tell this
gem exactly where it's schemas are (since it knows) rather than having
to rely on a slightly fragile "../publishing-api/content_schemas"
mapping which assumes Publishing API is in a specific directory.

Expected usage in Publishing API (or a similar Rails app that configured
it) would be:

```
GovukSchemas.content_schema_dir = Rails.root.join("content_schemas")
```

[1]: https://github.com/alphagov/publishing-api/pull/2184