Add missing regex generator for 1000-9999
A new regex was recently added to the specialist document schema [1]. As
this regex is not represented in this repo we see intermittent test
failures [2]. With this error:

```
  1) GovukIndex::ElasticsearchPresenter Specialist formats statutory instrument
     Failure/Error:
       example = GovukSchemas::RandomExample.for_schema(notification_schema: "specialist_document") do |payload|
         payload["details"]["metadata"] = metadata
         payload
       end

     RuntimeError:
                 Don't know how to generate random string for pattern "^[1-9][0-9]{3}$"

                 This propably means you've introduced a new regex in  publishing api.
                 Because it's very hard to generate a valid string from a regex alone,
                 we have to specify a method to generate random data for each regex in
                 the schemas.

                 To fix this:

                 - Add your regex to `lib/govuk_schemas/random.rb`
```

By adding this regex we can resolve this error.

I did wish to add tests for this change but this repo does not have a
convention for any testing of these regexs, so I quietly backed away
from it.

To test it locally I performed the following:

```
irb(main):005:0> 1000.times { |i| GovukSchemas::RandomExample.for_schema(frontend_schema: "specialist_document") }
=> 1000
```

Which prior to this change would exception.

[1]: https://github.com/alphagov/publishing-api/commit/613897340dd609c5d7540685a6f4d1d5be220c39#diff-e427ec772dc2597718b907f2db7772ad580d90452a76ce291114ddd0cfacb289R2191
[2]: https://github.com/alphagov/publishing-api/actions/runs/5013994959/jobs/8987739121