"One-of-everything" strategy for random generator
We're working on automatically generating the canonical GraphQL query
for a given content type. The method that we're currently developing
requires a content item as its input. Therefore we're after an example
content item that has every possible field populated.

Our existing randomly-generated content items will randomly leave out
optional fields and arrays without a sufficient `minItems` setting. This
new strategy offers a way to generate an otherwise random content item,
but with every possible field populated.

Note: there's no major reason that we've limited arrays' size to 1 as
opposed to only setting its minimum size to 1. I was drawn to the
particular name of the strategy first and then let that lead to this
logic.

Also note: I can't guarantee that I've covered all gaps in the random
generator here because I'm very new to this codebase.