Enable random `seed` to be supplied to #for_schema
This change is just wiring-up an existing feature of RandomExample's
initialize method to this alternative constructor.

The result is a bit unusual-looking because of the key:value shape of
the existing param, but this feels like the least convoluted way to
support the optional params alongside that required key:value pair.